+++ lib/index.ts
@@ -1,4 +1,4 @@
-export * from "./context";
+export * from "./form-context";
 export * from "./field";
 export * from "./form";
 export * from "./types";
